pyproject.toml
[tool.poetry]
name = "bonch-parser"
version = "0.1.0"
description = ""
authors = ["Daniil Kharkov <fadeddexofan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.6"
httpx = "^0.16.1"
pydantic = "^1.7.3"
fastapi = "^0.62.0"
starlette = "^0.13.6"
uvicorn = { version = "^0.13.1", extras = ["standard"] }
aio-pika = "^6.7.1"
tortoise-orm = "^0.16.18"
uvloop = "^0.14.0"
asyncpg = "^0.21.0"
aerich = "^0.4.2"
python-dotenv = "^0.15.0"
sentry-sdk = "^0.19"
pyppeteer = "^0.2.2"
celery = { version = "*", extras = ["redis"] }
aioboto3 = "^8.2.0"
aioredis = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
